+++ google/cloud/logging_v2/handlers/structured_log.py
@@ -22,27 +22,9 @@ import logging.handlers
 
 from google.cloud.logging_v2.handlers._helpers import format_stackdriver_json
 from google.cloud.logging_v2.handlers._helpers import get_request_data
+from google.cloud.logging_v2.handlers.handlers import CloudLoggingFilter
+from google.cloud.logging_v2.handlers.handlers import GCP_FORMAT
 
-GCP_FORMAT = '{"message": "%(message)s", "severity": "%(levelname)s", "logging.googleapis.com/trace": "%(trace)s", "logging.googleapis.com/sourceLocation": { "file": "%(filename)s", "line": "%(lineno)d", "function": "%(funcName)s"}, "httpRequest": {"requestMethod": "%(request_method)s", "requestUrl": "%(request_url)s", "userAgent": "%(user_agent)s", "protocol": "%(protocol)s"} }'
-
-
-class GCPFilter(logging.Filter):
-
-    def __init(self, project=None):
-        self.project = project
-
-    def filter(self, record):
-        inferred_http, inferred_trace = get_request_data()
-        if inferred_trace is not None and self.project is not None:
-            inferred_trace = f"projects/{self.project_id}/traces/{inferred_trace}"
-
-        record.trace = trace_id = record.trace or inferred_trace or ""
-        record.http_request = record.http_request or record.httpRequest or inferred_http or {}
-        record.request_method = record.http_request.get('requestMethod', "")
-        record.request_url = record.http_request.get('requestUrl', "")
-        record.user_agent = record.http_request.get('userAgent', "")
-        record.protocol = record.http_request.get('protocol', "")
-        return True
 
 class StructuredLogHandler(logging.StreamHandler):
     """Handler to format logs into the Cloud Logging structured log format,
@@ -60,7 +42,7 @@ class StructuredLogHandler(logging.StreamHandler):
         self.project_id = project
 
         # add extra keys to log record
-        self.addFilter(GCPFilter())
+        self.addFilter(CloudLoggingFilter(project))
 
         # make logs appear in GCP structured logging format
         self.formatter = logging.Formatter(GCP_FORMAT)
